Adding in first container code for award amounts
