+++ src/js/helpers/searchHelper.js
@@ -226,6 +226,21 @@ export const fetchAwardV2 = (awardId) => {
     };
 };
 
+export const fetchAwardsAmount = (awardId) => {
+    const source = CancelToken.source();
+    return {
+        promise: Axios.request({
+            url: `v2/awards/idvs/amounts/${awardId}/`,
+            baseURL: kGlobalConstants.API,
+            method: 'get',
+            cancelToken: source.token
+        }),
+        cancel() {
+            source.cancel();
+        }
+    };
+};
+
 // Fetch Individual Award's Transactions
 export const fetchAwardTransaction = (params) => {
     const source = CancelToken.source();
